Refactor checks
