+++ pythonwhat/check_syntax.py
@@ -1,4 +1,4 @@
-from pythonwhat.check_wrappers import scts
+from pythonwhat.checks.check_wrappers import scts
 from pythonwhat.State import State
 from pythonwhat.probe import Node, Probe, TEST_NAMES
 from pythonwhat.utils import include_v1
@@ -6,7 +6,6 @@ from pythonwhat import test_funcs
 from functools import partial, reduce, wraps
 import inspect
 import copy
-import os
 
 # TODO: could define scts for check_wrappers at the module level
 ATTR_SCTS = scts.copy()
