+++ pythonwhat/checks/has_funcs.py
