+++ pythonwhat/test_funcs/test_compound_statement.py
@@ -1,8 +1,8 @@
-from pythonwhat.check_funcs import check_node, check_part, check_part_index, with_context
+from pythonwhat.checks.check_funcs import check_node, check_part, check_part_index, with_context
 from pythonwhat.test_funcs.utils import fix_format, stringify, call
-from pythonwhat.check_logic import multi
-from pythonwhat.has_funcs import has_equal_part_len, has_equal_part, has_equal_value, has_equal_output
-from pythonwhat.check_has_context import has_context
+from pythonwhat.checks.check_logic import multi
+from pythonwhat.checks.has_funcs import has_equal_part_len, has_equal_part, has_equal_value, has_equal_output
+from pythonwhat.checks.check_has_context import has_context
 from functools import partial
 from pythonwhat import utils
