+++ tests/test_v2_only.py
@@ -1,5 +1,4 @@
 import pytest
-import os
 import helper
 import importlib
